add 最爱and fix bug

+++ XZMusic/Views/ViewController/PlayingViewController/View/XZMusicLrcView.h
@@ -11,5 +11,5 @@
 @interface XZMusicLrcView : UIView
 
 - (void)initLrcViewWithPath:(NSString *)lrcPath;
-- (void)moveLrcWithTime:(int)time;
+- (void)moveLrcWithTime:(NSInteger)time;
 @end
